working on handling face disapearing
I started working on handling the face of the target disappearing or moving out of frame. In a nutshell, I am stopping the timeout, resetting the face detection variables and sending an error message to the front end.

It work, but slow for some reason (maybe because it block?) - anyhow need to fix.

Also, re-auth does not work afterwords, I will look into that too.