Cleaned up model
I made put the CNN creation all into one file. Its not done yet, and has to be tested. Mainly, I need to figure out how to do a training method, and also see if this code works. The purpose of this commit is to set a general idea of how the model behaves, and the flow of data from layer to layer.

The idea is: we initialise the CNN and a dense layer, and then piece together the CNN as well as the last Dense layer in the call() method - this allows Keras to understadn that the CNN and Dense layer are modules, but both part of the overall model.